% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_functions.R
\docType{package}
\name{androidpublisher_googleAuthR}
\alias{androidpublisher_googleAuthR}
\alias{androidpublisher_googleAuthR-package}
\title{Google Play Developer API
Lets Android application developers access their Google Play accounts.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:35:20
filename: /Users/mark/dev/R/autoGoogleAPI/googleandroidpublisherv11.auto/R/androidpublisher_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/androidpublisher
}
}

