% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{SubscriptionPurchase}
\alias{SubscriptionPurchase}
\title{SubscriptionPurchase Object}
\usage{
SubscriptionPurchase(autoRenewing = NULL, initiationTimestampMsec = NULL,
  validUntilTimestampMsec = NULL)
}
\arguments{
\item{autoRenewing}{Whether the subscription will automatically be renewed when it reaches its current expiry time}

\item{initiationTimestampMsec}{Time at which the subscription was granted, in milliseconds since the Epoch}

\item{validUntilTimestampMsec}{Time at which the subscription will expire, in milliseconds since the Epoch}
}
\value{
SubscriptionPurchase object
}
\description{
SubscriptionPurchase Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A SubscriptionPurchase resource indicates the status of a user's subscription purchase.
}
\seealso{
Other res: \code{\link{InappPurchase}}
}

